requirements.txt
streamlit>=1.29.0
faiss-cpu>=1.7.4
sentence-transformers>=2.7.0
transformers>=4.44.0
torch>=2.0.0
numpy>=1.23.0
openai>=1.40.0
